+++ src/network/mod.rs
@@ -1,12 +1,12 @@
-//! Core engine components
+//! Core network engine components
 //!
 //! This module contains the core components that power the network testing engine.
 
-pub mod network;
+use pnet::datalink::{self, NetworkInterface};
+
 pub mod worker;
 pub mod simulation;
 pub mod target;
-pub mod traits;
 pub mod worker_manager;
 
 // Re-export commonly used types
@@ -14,7 +14,19 @@ pub use simulation::{Simulation, SimulationRAII};
 pub use target::MultiPortTarget;
 pub use worker_manager::Workers;
 
-// Re-export core traits from consolidated module
-pub use traits::{
-    StatsCollector, PacketBuilder as PacketBuilderTrait, TargetProvider, WorkerConfig
-};
+// Network interface utilities
+pub fn list_network_interfaces() -> Vec<NetworkInterface> {
+    datalink::interfaces()
+}
+
+pub fn find_interface_by_name(name: &str) -> Option<NetworkInterface> {
+    datalink::interfaces()
+        .into_iter()
+        .find(|iface| iface.name == name)
+}
+
+pub fn default_interface() -> Option<NetworkInterface> {
+    datalink::interfaces()
+        .into_iter()
+        .find(|iface| iface.is_up() && !iface.is_loopback() && !iface.ips.is_empty())
+}
